+++ lending/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v0.1.0] lending
 ### Releases
 - [PyPI v0.1.0] https://pypi.org/project/codat-lending/0.1.0 - lending
+
+## 2023-09-13 11:13:49
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.82.5 (2.108.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.0.0] lending
+### Releases
+- [PyPI v2.0.0] https://pypi.org/project/codat-lending/2.0.0 - lending